* commit '8f9d76e4190795ce8ab51956a39caa236e045cb2 [formerly 221fe62c8f66e635a3f1ad98bf2cb754a0480fd5]': (34 commits)
  INFUND-7367 Update jira number on todo comments
  INFUND-7367 Update jira number on todo comments
  INFUND-7367 - Addresses PR comments
  INFUND-7367 - Updates TODOs in templates/populators
  INFUND-7696 - Fixes failing Robot test on Application overview page
  updated view template to match latest in prot
  INFUND-7367 - Fixes failing robot tests on Innovation area column
  INFUND-7369 - Adds additional unit tests for query param preservation in all/submitted application dashboards
  INFUND-7371 - Fixes innovation area typo on Submitted Applications dashboard
  INFUND-7369 - Adds additional check in ApplicationSummaryMapperTest for getting correct numberOfPartners when there is an assessor process role
  INFUND-7369 - Fixes ApplicationSummaryMapper not settings ApplicationSummaryResource#numberOfPartners correctly
  INFUND-7369 - Fixes Innovation area typo
  INFUND-7655 - Adds back link origin for Funding applications dashboard
  INFUND-7696 - Removes unused block from CompetitionManagementApplicationController
  INFUND-7655 - Removes cookie-based back button and replaces with query parameters with origin enum
  INFUND-7367 - Fixes additional unit test failures from changing longs to ints in Application/CompetitionInviteStatistics
  INFUND-7367 - Removes typo from CompetitionInviteControllerIntegrationTest
  INFUND-7367
  updated to styling and view templates
  INFUND-7367 - Refactors ApplicationStatistics/CompetitionInviteStatistics to use ints rather than longs
  ...


Former-commit-id: 6c5f8389d15670625b4d63675934a5e4ad481428

+++ ifs-resources/src/main/java/org/innovateuk/ifs/competition/resource/CompetitionInAssessmentKeyStatisticsResource.java
@@ -5,49 +5,49 @@ import org.apache.commons.lang3.builder.HashCodeBuilder;
 
 public class CompetitionInAssessmentKeyStatisticsResource {
 
-    private long assignmentCount;
-    private long assignmentsWaiting;
-    private long assignmentsAccepted;
-    private long assessmentsStarted;
-    private long assessmentsSubmitted;
+    private int assignmentCount;
+    private int assignmentsWaiting;
+    private int assignmentsAccepted;
+    private int assessmentsStarted;
+    private int assessmentsSubmitted;
 
-    public long getAssignmentCount() {
+    public int getAssignmentCount() {
         return assignmentCount;
     }
 
-    public void setAssignmentCount(long assignmentCount) {
+    public void setAssignmentCount(int assignmentCount) {
         this.assignmentCount = assignmentCount;
     }
 
-    public long getAssignmentsWaiting() {
+    public int getAssignmentsWaiting() {
         return assignmentsWaiting;
     }
 
-    public void setAssignmentsWaiting(long assignmentsWaiting) {
+    public void setAssignmentsWaiting(int assignmentsWaiting) {
         this.assignmentsWaiting = assignmentsWaiting;
     }
 
-    public long getAssignmentsAccepted() {
+    public int getAssignmentsAccepted() {
         return assignmentsAccepted;
     }
 
-    public void setAssignmentsAccepted(long assignmentsAccepted) {
+    public void setAssignmentsAccepted(int assignmentsAccepted) {
         this.assignmentsAccepted = assignmentsAccepted;
     }
 
-    public long getAssessmentsStarted() {
+    public int getAssessmentsStarted() {
         return assessmentsStarted;
     }
 
-    public void setAssessmentsStarted(long assessmentsStarted) {
+    public void setAssessmentsStarted(int assessmentsStarted) {
         this.assessmentsStarted = assessmentsStarted;
     }
 
-    public long getAssessmentsSubmitted() {
+    public int getAssessmentsSubmitted() {
         return assessmentsSubmitted;
     }
 
-    public void setAssessmentsSubmitted(long assessmentsSubmitted) {
+    public void setAssessmentsSubmitted(int assessmentsSubmitted) {
         this.assessmentsSubmitted = assessmentsSubmitted;
     }
